* Mono.UIAutomation.Winforms/ToolStripProvider.cs: Give AccessibleName preedence over Text.
* ToolStripProviderTest.cs: Test that AccessibleName is used in
preference to Text for the name.

* Element.cs, ExpandCollapseSource.cs: Use child accessible for
ExpandCollapsePattern in a DataItem.

* Element.cs: Mark IsControlElement and IsContentElement as supported.

* Element.cs: Do not expose Grid/GridItem for a Tree/TreeItem.
Do not expose InvokePattern for a CheckBox.
Do not expose SelectionPattern for a MenuBar.

* UIAutomationClientTests/TableHierarchyTest.cs, GtkForm/GtkForm.cs:
Expect check boxes in the table for Winforms.
Fix name of boolean cells to be the same in Gtk as in Winforms.
Do not assume that a treeItem has no children when collapsed.
Test TreeView patterns.

* UIAutomationClientTests/AutomationElementTest.cs:
Assume that IsContentElement is true for a label in gtk.
Check that a Spinner does not expose InvokePattern.

* UIAutomationClientTests/AutomationElementTest.cs, GtkForm/GtkForm.cs:
Test checkBox patterns.

* UIAutomationClientTests/BaseTest.cs,
UIAutomationClientTests/TogglePatternTest.cs: Fix checkboxes.

* UIAutomationClientTests/BaseTest.cs: Catch
InvalidOperationException and display pattern name in VerifyPatterns.

* UIAutomationClientTests/MenuTest.cs: Test MenuBar patterns.

* AtspiUiaClientTests/TableHierarchyTest.cs, GtkForm/GtkForm.cs:
Test DataItem patterns.


svn path=/trunk/uia2atk/; revision=156800